Created Processor to process annotated classes;
The RepositoryProcessor class will process the annotated classes with
the RepositoryProperty annotation and generate a Repository class for
the this classes.

The generated classes will reside inside
target/generated-sources/annotations folder. The package name will be a
variant of the given class, changing the last name to "repository", and
the class name will be the annotated class name appended by the
"Repository" word;